Improves code functionality and quality. Search now supports paratial query matching, partial match results list (creates list of close matches to the query). Also cleaned up the UI and made the user messages clearer.
Also implements unity events that can be used with other code for routing.
Search button will now show possible matches from the list (based on partial text) for when user is still typing or unsure of full room name.
Go button confirms an exact room name and triggers the routing event to be used in other code. Each room has a destination ID that will serve as a key for routing.